+++ ADswaggerAPI/Models/Sales.cs
@@ -1,6 +1,6 @@
 public class Sales{
     //public int ID { get; set; }
     public string ArticleNumber { get; set; }
-    public int SalesPrice { get; set; }
+    public double SalesPrice { get; set; }
     public string Day { get; set; }
 }
