+++ ADswaggerAPI/Models/TotalDailyRevenue.cs
@@ -1,9 +1,9 @@
 public class TotalDailyRevenue{
     public string Day { get; set; }
-    public int Revenue { get; set; }
+    public double Revenue { get; set; }
     
     // Default Construction
-    public TotalDailyRevenue(string _day, int _revenue)
+    public TotalDailyRevenue(string _day, double _revenue)
     {
          Day = _day;
          Revenue = _revenue;
